% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_functions.R
\name{map_performance}
\alias{map_performance}
\title{Performance curve data frame}
\usage{
map_performance(x = seq(0, 1, length.out = 100), par_df)
}
\arguments{
\item{x}{A [0,1] vector of values to evaluate performance over}

\item{par_df}{A data frame produced by perform_df().}
}
\value{
A tidy data frame, containing a points along the environmental axis, and corresponding points for the performance axis, for each group and posterior draw input.
}
\description{
Constructs a tidy data frame representing the full perforance curve (facilitates plotting).
}
