% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_functions.R
\name{map_performance_fixed}
\alias{map_performance_fixed}
\title{Performance curve data frame}
\usage{
map_performance_fixed(x, par_df)
}
\arguments{
\item{x}{A vector of values to evaluate performance over.}

\item{par_df}{A data frame produced by perform_df().}
}
\value{
A tidy data frame, containing a points along the environmental axis, and corresponding points for the performance axis, for each group and posterior draw input.
}
\description{
Similar to map_performance(), but over a set of fixed points along the axis rather than being sampled relative to each species performance limits.
}
