% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_functions.R
\name{perform_df}
\alias{perform_df}
\title{Performance data frame}
\usage{
perform_df(stan_out, species_order)
}
\arguments{
\item{stan_out}{The name of the object output of stan_performance().}

\item{species_order}{A character vector of IDs that match the output dimensions of each parameter matrix.}
}
\value{
A tidy data frame, where each row is a draw from one of the input groups and each column is a model parameter. Additionally, several derived parameters, including the optimum, area, breadth, and area scaled by breadth (called special).
}
\description{
Constructs a tidy data frame from the output of stan_performance()
}
