% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_mu.R
\name{performance_mu}
\alias{performance_mu}
\title{Performance curve}
\usage{
performance_mu(xs, shape1, shape2, stretch, x_min, x_max)
}
\arguments{
\item{xs}{Vector of environmental axis values}

\item{shape1}{First parameter controlling asymmetry along environmental axis.}

\item{shape2}{Second parameter controlling asymmetry along environmental axis.}

\item{stretch}{Parameter controlling stretch along the performance axis.}

\item{x_min}{Parameter controlling where perforamance falls to zero below optimum.}

\item{x_max}{Parameter to controlling where perforamance falls to zero above optimum.}
}
\value{
Vector of performance values.
}
\description{
Deterministic function representing an individual performance/tolerance curve.
}
\examples{
curve(performance_mu(xs = x, 2, 3, 1, 1, 5), from = 1, to = 5)
}
