% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_functions.R
\name{posterior_quantile}
\alias{posterior_quantile}
\title{Generate psuedo-observed quantiles from performance curve parameters}
\usage{
posterior_quantile(x, p, par_df)
}
\arguments{
\item{x}{A vector of values to evaluate performance over}

\item{p}{Vector of probability values passed to qnorm -- the amount of probability density right of the returned quantiles}

\item{par_df}{A data frame like the one produced by perform_df(). MUST contain columns named: draw, species, x_min, x_max, shape1, shape2, stretch, and nu.}
}
\value{
A tidy data frame, containing upper and lower quantiles along the environmental axis, and corresponding points for the performance axis, for each group and posterior draw input.
}
\description{
Constructs a tidy data frame of generated data given a set of input parameters.
}
