% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_functions.R
\name{predictions}
\alias{predictions}
\title{Prediction quantiles}
\usage{
predictions(x, spp, par_df, x_draws, p)
}
\arguments{
\item{x}{A vector of values to evaluate performance over}

\item{spp}{The species to produce predictions over, one at a time is recommended.}

\item{par_df}{A data frame like the one produced by perform_df(). MUST contain columns named: draw, species, x_min, x_max, shape1, shape2, stretch, and nu.
c p Vector of probability values passed to qnorm -- the amount of probability density right of the returned quantiles}
}
\value{
A tidy data frame, containing averaged upper and lower quantiles along the environmental axis, and corresponding points for the performance axis.
}
\description{
Generate performance prediction quantiles over multiple posterior draws. Good for visualzation.
}
