% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_performance.R
\name{stan_performance}
\alias{stan_performance}
\title{Performance curve inference with Stan}
\usage{
stan_performance(df, response, treatment, group_ids, file_id = NULL,
  max_treedepth = 10, iter = 2000, thin = 1, adapt_delta = 0.95,
  chains = 4, shape1_pr_mu = 4, shape1_pr_sig = 1, shape2_pr_mu = 4,
  shape2_pr_sig = 1, stretch_pr_mu = 0, stretch_pr_sig = 1,
  nu_pr_shape = 2, nu_pr_scale = 3, min_pr_mu, min_pr_sig = 1, max_pr_mu,
  max_pr_sig = 1, ...)
}
\arguments{
\item{df}{Data frame containing response, treatment, and groups of interest.}

\item{response}{the unquoted column name containing the response trait (zero and positive real numbers expected).}

\item{treatment}{the unquoted column name containing the treatment values (real values expected).}

\item{group_ids}{the unquoted column name containing the fixed groups (i.e. species, populations, etc.).}

\item{file_id}{Optional file prefix to write Stan samples to (stan() sample_file argument).}

\item{...}{Further arugments to pass to Stan's sampling() function.}
}
\description{
Runs performance/tolerance curve model
}
\details{
This function provides the interface to Stan. Arguments ending in "pr_mu" and  "pr_sig" are the mean and standard deviations of normal prior distributions.
}
