% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_functions.R
\name{posterior_predict}
\alias{posterior_predict}
\title{Generate perforamcne curves data from parameters}
\usage{
posterior_predict(x, shape1, shape2, stretch, x_min, x_max, nu, species_id)
}
\arguments{
\item{x}{A vector of values to evaluate performance over}

\item{shape1}{First parameter controlling asymmetry along environmental axis.}

\item{shape2}{Second parameter controlling asymmetry along environmental axis.}

\item{stretch}{Parameter controlling stretch along the performance axis.}

\item{x_min}{Parameter controlling where perforamance falls to zero below optimum.}

\item{x_max}{Parameter to controlling where perforamance falls to zero above optimum.}

\item{species_id}{Character string to connect to returned rows.}
}
\value{
A tidy data frame, containing a points along the environmental axis, and corresponding points for the performance axis, for each group and posterior draw input.
}
\description{
Constructs a tidy data frame of generated data given a set of input parameters.
}
